+++ example/integration_test/native_marionette_support.dart
@@ -1,5 +1,4 @@
 import 'dart:convert';
-import 'dart:typed_data';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter_ai_communications/flutter_ai_communications.dart';