+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -5,6 +5,8 @@
 import FlutterMacOS
 import Foundation
 
+import flutter_ai_communications_macos
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
+  FlutterAiCommunicationsPlugin.register(with: registry.registrar(forPlugin: "FlutterAiCommunicationsPlugin"))
 }